Move `can_use` method out of `fontbakery.utils`
