+++ tests/test_utils.py
@@ -2,13 +2,10 @@ import pytest
 
 from fontbakery.utils import (
     bullet_list,
-    can_shape,
     pretty_print_list,
     text_flow,
     unindent_and_unwrap_rationale,
 )
-from fontTools.ttLib import TTFont
-from fontbakery.codetesting import portable_path
 
 
 def test_text_flow():
@@ -89,14 +86,6 @@ def test_text_flow():
 #                     space_padding=True) == (     " One   \n"
 #                                             "      Two   \n"
 #                                             "      Three ")
-
-
-def test_can_shape():
-    font = TTFont(portable_path(
-        "data/test/source-sans-pro/OTF/SourceSansPro-Regular.otf"
-    ))
-    assert can_shape(font, "ABC")
-    assert not can_shape(font, "こんにちは")
 
 
 def test_unindent_and_unwrap_rationale():
